integer64 should be long long int, not long int
Compiler warning on linux cluster:

In file included from utils.c:1:0:
utils.c: In function ‘asCharacterInteger64’:
data.table.h:50:30: warning: format ‘%lli’ expects argument of type ‘long long int’, but argument 4 has type ‘int64_t {aka long int}’ [-Wformat=]
 #define INTEGER64_ASCHAR_FMT "%lli"
                              ^
utils.c:377:44: note: in expansion of macro ‘INTEGER64_ASCHAR_FMT’
       snprintf(buff, INTEGER64_ASCHAR_LEN, INTEGER64_ASCHAR_FMT, px[i]);
                                            ^~~~~~~~~~~~~~~~~~~~

+++ src/utils.c
@@ -361,7 +361,7 @@ SEXP coerceUtf8IfNeeded(SEXP x) {
 SEXP asCharacterInteger64(SEXP x) {
   const int64_t lenx = xlength(x);
   SEXP coerced = PROTECT(allocVector(STRSXP, lenx));
-  int64_t * px = (int64_t *) REAL(x);
+  long long int *px = (long long int *) REAL(x);
   static char buff[INTEGER64_ASCHAR_LEN];
   for (int64_t i=0; i<lenx; ++i) {
     if (px[i]==NA_INTEGER64) {
